1. changed all uint with uint16_t (uint is not standard c++ type, types defined in <cstdint>) 2. changed in gmm.hh const uint ngaus=2; which is used for init an array. this in not allowed because the class is not "const" (checked also with http://coliru.stacked-crooked.com). modified in static const uint ngaus=2; 3. added selection on passing through tracks for performing calibration 4. added parameters to equalization_template.steer

+++ EventReconstruction/Calo/src/LCaloCalibration.cc
@@ -120,7 +120,7 @@ LCaloCalibration& LCaloCalibration::operator+=(const LCaloCalibration& rhs) // c
   std::vector<double> skew(NPMT,0.);
   std::vector<double> kurt(NPMT,0.);
 
-  for(uint ipmt=0; ipmt<NPMT; ++ipmt) {
+  for(uint16_t ipmt=0; ipmt<NPMT; ++ipmt) {
     double tmp = rhs.GetPedestal()[ipmt];
     ped[ipmt] = pedestal[ipmt]+tmp;
     double tmpsig = rhs.GetSigma()[ipmt];
@@ -135,7 +135,7 @@ LCaloCalibration& LCaloCalibration::operator+=(const LCaloCalibration& rhs) // c
        (sigma[ipmt]*sigma[ipmt]*sigma[ipmt]*sigma[ipmt]+tmpsig*tmpsig*tmpsig*tmpsig);  // correlations neglected!!!
     */
   }
-  for(uint ipmt=0; ipmt<NPMT; ++ipmt) {
+  for(uint16_t ipmt=0; ipmt<NPMT; ++ipmt) {
     pedestal[ipmt] = ped.at(ipmt);
     sigma[ipmt] = sqrt(var.at(ipmt));
     /*  outliers[ipmt] = outl[ipmt];
